add separate renderer
