+++ app/views/hyrax/articles/_attribute_rows.html.erb
@@ -1,7 +1,7 @@
 <%= presenter.attribute_to_html(:alternative_title, html_dl: true, label: "Alternate title") %>
 <%= presenter.attribute_to_html(:creator_display, label: 'Creator', render_as: :person, html_dl: true) %>
 <%= presenter.attribute_to_html(:translator_display, label: 'Translator', render_as: :person, html_dl: true) %>
-<%= presenter.attribute_to_html(:abstract, html_dl: true) %>
+<%= presenter.attribute_to_html(:abstract, render_as: :formatted_text, html_dl: true) %>
 <%= presenter.attribute_to_html(:date_issued, html_dl: true, label: "Date of publication") %>
 <%= presenter.attribute_to_html(:keyword, render_as: :faceted, html_dl: true) %>
 <%= presenter.attribute_to_html(:subject, render_as: :faceted, html_dl: true) %>
